refactor: move io methods to FileIO
