+++ bundle-utils/src/Utils.h
@@ -92,40 +92,6 @@ namespace jetpack::utils {
         return ms.count();
     }
 
-    inline std::string To_UTF8(const std::u16string &s) {
-#ifndef _WIN32
-        std::wstring_convert<std::codecvt_utf8<char16_t>, char16_t> conv;
-        return conv.to_bytes(s);
-#else
-        std::size_t buffer_size = s.size() * 2;
-        char* buffer = new char[buffer_size];
-        memset(buffer, 0, buffer_size);
-
-        int ret = WideCharToMultiByte(CP_UTF8, WC_COMPOSITECHECK,
-            reinterpret_cast<const wchar_t*>(s.c_str()),
-            s.size(), buffer, buffer_size, NULL, NULL);
-        
-        if (ret == 0) {
-            delete[] buffer;
-            return "";
-        }
-
-        delete[] buffer;
-        return std::string(buffer, ret);
-#endif
-    }
-
-    inline bool IsFileExist(const std::string& path) {
-#ifndef _WIN32
-        return access(path.c_str(), F_OK) == 0;
-#else
-        DWORD dwAttrib = GetFileAttributesA(path.c_str());
-
-        return (dwAttrib != INVALID_FILE_ATTRIBUTES &&
-            !(dwAttrib & FILE_ATTRIBUTE_DIRECTORY));
-#endif
-    }
-
 }
 
 template <typename T>
